+++ lib/stellar/effects.ex
@@ -39,4 +39,26 @@ defmodule Stellar.Effects do
     query = Base.process_query_params(params)
     Base.get("/accounts/#{accountId}/effects#{query}")
   end
+
+  @doc """
+  Returns all effects for given operation
+
+  See `all/1` for allowed optional params
+  """
+  @spec all_for_operation(binary, Keyword.t) :: {Stellar.status, map}
+  def all_for_operation(operationId, params \\ []) do
+    query = Base.process_query_params(params)
+    Base.get("/operations/#{operationId}/effects#{query}")
+  end
+
+  @doc """
+  Returns all effects for given ledger
+
+  See `all/1` for allowed optional params
+  """
+  @spec all_for_ledger(binary, Keyword.t) :: {Stellar.status, map}
+  def all_for_ledger(ledgerId, params \\ []) do
+    query = Base.process_query_params(params)
+    Base.get("/ledgers/#{ledgerId}/effects#{query}")
+  end
 end