+++ lib/stellar/payments.ex
@@ -31,4 +31,15 @@ defmodule Stellar.Payments do
     query = Base.process_query_params(params)
     Base.get("/accounts/#{accountId}/payments#{query}")
   end
+
+  @doc """
+  Returns all payments for given ledger
+
+  See `all/1` for allowed optional params
+  """
+  @spec all_for_ledger(binary, Keyword.t) :: {Stellar.status, map}
+  def all_for_ledger(ledgerId, params \\ []) do
+    query = Base.process_query_params(params)
+    Base.get("/ledgers/#{ledgerId}/payments#{query}")
+  end
 end