+++ test/stellar/operations_test.exs
@@ -25,4 +25,12 @@ defmodule Stellar.Operations.Test do
 
     assert {:ok, %{"_embedded" => _}} = Operations.all_for_account("123456")
   end
+
+  test "get all operations for a ledger", %{bypass: bypass} do
+    Bypass.expect_once bypass, "GET", "/ledgers/123456/operations", fn conn ->
+      Plug.Conn.resp(conn, 200, ~s<{"_embedded": { "records": [] }}>)
+    end
+
+    assert {:ok, %{"_embedded" => _}} = Operations.all_for_ledger("123456")
+  end
 end